add login

+++ Api/ProjectController.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using ProjectManage.Data;
@@ -8,79 +10,72 @@ using ProjectManage.Models;
 
 namespace ProjectManage.Controllers
 {
+    [Authorize]
     [Route("api/project")]
     public class ProjectController : ControllerBase
     {
         private readonly ApplicationDbContext _context;
-
-
         public ProjectController(ApplicationDbContext context)
         {
             _context = context;
         }
-
         // getprojct by name: /api/Project/1
-
         [HttpGet("{id}")]
         public async Task<ActionResult<Project>> Get(int id)
         {
-            var project = await _context.Projects.FindAsync(id);
-            if (project == null)
-                return BadRequest();
-            return project;
+            // var project = await _context.Projects.FindAsync(id);
+            var userID =Int32.Parse(User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value);
+            var newQuery = _context.Projects.Where(p => p.UserProjects.Any(pu => pu.UserId == userID && pu.ProjectId == id))
+                .Select(p => new
+                {
+                    id = p.Id,
+                    name = p.Name,
+                    created = p.CreatedAt,
+                    updated = p.UpdatedAt,
+                    manager = p.ManagerId,
+                    status = p.Status,
+                    thumb = p.Thumb,
+                    users = p.UserProjects.Select(du => new
+                    {
+                        userId = du.User.Id,
+                        fullName = du.User.FullName,
+                        email = du.User.Email,
+                        avatar = du.User.Avatar,
+                    }),
+                });
+                var dataNeW = await newQuery.ToListAsync();
+                return Ok(dataNeW);
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll([FromQuery] string search = null, int? userId = null)
+        public async Task<IActionResult> GetAll()
         {
-
-            var query = _context.Projects.AsQueryable();
-
-            if (!string.IsNullOrWhiteSpace(search))
-            {
-                query = query.Where(item => item.Name.Contains(search));
-            }
-
-            if (userId.HasValue)
-            {
-                var newQuery = _context.Projects.Where(p => p.UserProjects.Any(pu => pu.UserId == userId))
-                     .Select(p => new
-                     {
-                         id = p.Id,
-                         name = p.Name,
-                         created = p.CreatedAt,
-                         updated = p.UpdatedAt,
-                         manager = p.ManagerId,
-                         status = p.Status,
-                         thumb = p.Thumb,
-                         users = p.UserProjects.Select(du => new
-                         {
-                             userId = du.User.Id,
-                             fullName = du.User.FullName,
-                             email = du.User.Email,
-                             avatar = du.User.Avatar,
-                         }),
-
-                     });
-
+            var userID =Int32.Parse(User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value);
+            var newQuery = _context.Projects.Where(p => p.UserProjects.Any(pu => pu.UserId == userID))
+                .Select(p => new
+                {
+                    id = p.Id,
+                    name = p.Name,
+                    created = p.CreatedAt,
+                    updated = p.UpdatedAt,
+                    manager = p.ManagerId,
+                    status = p.Status,
+                    thumb = p.Thumb,
+                    users = p.UserProjects.Select(du => new
+                    {
+                        userId = du.User.Id,
+                        fullName = du.User.FullName,
+                        email = du.User.Email,
+                        avatar = du.User.Avatar,
+                    }),
+                });
                 var dataNeW = await newQuery.ToListAsync();
                 return Ok(dataNeW);
-
-            }
-
-
-            var data = await query.ToListAsync();
-
-            return Ok(data);
         }
-
-
         // post-create new Project: api/Project
         [HttpPost]
         public async Task<ActionResult<Project>> Create([FromBody] Project model, ProjectViewModel viewModel)
         {
-
-
             if (!ModelState.IsValid)
             {
                 var errors = ModelState.Select(x => x.Value.Errors)
@@ -88,11 +83,9 @@ namespace ProjectManage.Controllers
                                     .ToList();
                 return BadRequest(errors);
             }
-
             /*==============================
             Insert table Projects
             ==============================*/
-
             model.CreatedAt = DateTime.Now;
             model.UpdatedAt = DateTime.Now;
             model.Status = ProjectStatus.Pending;
@@ -138,6 +131,9 @@ namespace ProjectManage.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> Edit(int id, [FromBody] Project model, ProjectViewModel viewModel)
         {
+            var userID =Int32.Parse(User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value);
+             // return Ok(found);
+            
             if (!ModelState.IsValid)
             {
                 var errors = ModelState.Select(x => x.Value.Errors)
@@ -145,13 +141,11 @@ namespace ProjectManage.Controllers
                                 .ToList();
                 return BadRequest(errors);
             }
-
-
-            var found = await _context.Projects.FindAsync(id);
-
+            // var found = await _context.Projects.FindAsync(id);
+            var found = await _context.Projects.FirstAsync(p => p.UserProjects.Any(pu => pu.UserId == userID && pu.ProjectId == id));
+          
             if (found != null)
             {
-
                 /*==============================
                     Update table Projects
                     ==============================*/
@@ -167,7 +161,6 @@ namespace ProjectManage.Controllers
                     1. Delete
                     2. Insert
                     ==============================*/
-
                 /* =========== 1 ==========*/
 
                 var query = _context.UserProjects.AsQueryable();
@@ -181,18 +174,12 @@ namespace ProjectManage.Controllers
                     await _context.SaveChangesAsync();
 
                 }
-                 
-                
                 /* =========== 2 ==========*/
-
                 if (model.UserProjects.Count > 0)
                 {
                     await _context.UserProjects.AddRangeAsync(model.UserProjects);
                     await _context.SaveChangesAsync();
                 }
-
-
-
                 /*==============================
                 Insert table ProjectsHistories
                 ==============================*/
@@ -224,26 +211,24 @@ namespace ProjectManage.Controllers
                 viewModel.FullName = FullName;
                 viewModel.Content = history.Content;
                 viewModel.CreatedAt = history.CreatedAt;
-
                 return Ok(viewModel);
-
             }
-
             return NotFound("Không tồn tại dự án");
         }
-
 
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var userID =Int32.Parse(User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value);
+         
             var found = await _context.Projects.FindAsync(id);
-            if (found != null)
+            if (found != null && userID == found.ManagerId)
             {
                 _context.Projects.Remove(found);
                 await _context.SaveChangesAsync();
                 return Ok(new { success = "Xóa dự án thành công" });
             }
-            return NotFound();
+            return NotFound("bạn ko phải quản lý của dự án");
         }
 
     }
